[#98] Convert the ToStringBuilder/ToStringContext from Xtend to Java.
- Remove xtend-gen folder from the classpath.
